+++ crates/aviutl2/src/lib.rs
@@ -26,6 +26,9 @@
 pub use aviutl2_sys as sys;
 pub use log;
 
+#[doc(inline)]
+pub use aviutl2_macros::plugin;
+
 #[cfg(feature = "input")]
 pub mod input;
 
@@ -35,7 +38,20 @@ pub mod output;
 #[cfg(feature = "filter")]
 pub mod filter;
 
+#[doc(hidden)]
+#[path = "internal_base.rs"]
+pub mod __internal_base;
+
 pub mod common;
 pub mod utils;
 pub use anyhow;
 pub use common::*;
+
+#[macro_export]
+macro_rules! internal_module {
+    ($($item:item)*) => {
+        const _: () = {
+            $($item)*
+        };
+    }
+}